requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.9.2
anyio=4.6.2.post1=pyhd8ed1ab_0
appnope=0.1.4=pyhd8ed1ab_1
argon2-cffi=23.1.0=pyhd8ed1ab_1
argon2-cffi-bindings=21.2.0=py312h024a12e_5
arrow=1.3.0=pyhd8ed1ab_0
asttokens=3.0.0=pyhd8ed1ab_1
async-lru=2.0.4=pyhd8ed1ab_0
attrs=24.2.0=pyh71513ae_0
babel=2.16.0=pyhd8ed1ab_1
beautifulsoup4=4.12.3=pyha770c72_1
bleach=6.2.0=pyhd8ed1ab_1
brotli=1.1.0=hd74edd7_2
brotli-bin=1.1.0=hd74edd7_2
brotli-python=1.1.0=py312hde4cb15_2
bzip2=1.0.8=h99b78c6_7
ca-certificates=2024.8.30=hf0a4a13_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2024.8.30=pyhd8ed1ab_0
cffi=1.17.1=py312h0fad829_0
charset-normalizer=3.4.0=pyhd8ed1ab_1
colorama=0.4.6=pyhd8ed1ab_1
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.3.1=py312hb23fbb9_0
cpython=3.12.8=py312hd8ed1ab_1
cycler=0.12.1=pyhd8ed1ab_1
debugpy=1.8.9=py312hd8f9ff3_0
decorator=5.1.1=pyhd8ed1ab_1
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_1
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_0
filelock=3.16.1=pyhd8ed1ab_1
fonttools=4.55.2=py312h998013c_0
fqdn=1.5.1=pyhd8ed1ab_1
freetype=2.12.1=hadb7bae_2
fsspec=2024.10.0=pyhff2d567_0
giflib=5.2.2=h93a5062_0
gmp=6.3.0=h7bae524_2
gmpy2=2.1.5=py312h524cf62_3
h11=0.14.0=pyhd8ed1ab_1
h2=4.1.0=pyhd8ed1ab_1
hpack=4.0.0=pyhd8ed1ab_1
httpcore=1.0.7=pyh29332c3_1
httpx=0.28.0=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd8ed1ab_1
idna=3.10=pyhd8ed1ab_1
importlib-metadata=8.5.0=pyha770c72_1
importlib_resources=6.4.5=pyhd8ed1ab_1
ipykernel=6.29.5=pyh57ce528_0
ipython=8.30.0=pyh707e725_0
ipywidgets=8.1.5=pyhd8ed1ab_0
isoduration=20.11.0=pyhd8ed1ab_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.4=pyhd8ed1ab_1
joblib=1.4.2=pyhd8ed1ab_0
json5=0.10.0=pyhd8ed1ab_1
jsonpointer=3.0.0=py312h81bd7bf_1
jsonschema=4.23.0=pyhd8ed1ab_0
jsonschema-specifications=2024.10.1=pyhd8ed1ab_0
jsonschema-with-format-nongpl=4.23.0=hd8ed1ab_0
jupyter=1.1.1=pyhd8ed1ab_0
jupyter-lsp=2.2.5=pyhd8ed1ab_0
jupyter_client=8.6.3=pyhd8ed1ab_0
jupyter_console=6.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
jupyter_events=0.10.0=pyhd8ed1ab_0
jupyter_server=2.14.2=pyhd8ed1ab_0
jupyter_server_terminals=0.5.3=pyhd8ed1ab_0
jupyterlab=4.3.2=pyhd8ed1ab_0
jupyterlab_pygments=0.3.0=pyhd8ed1ab_2
jupyterlab_server=2.27.3=pyhd8ed1ab_0
jupyterlab_widgets=3.0.13=pyhd8ed1ab_0
kiwisolver=1.4.7=py312h6142ec9_0
krb5=1.21.3=h237132a_0
lcms2=2.16=ha0e7c42_0
lerc=4.0.0=h9a09cb3_0
libabseil=20240722.0=cxx17_hf9b8971_1
libblas=3.9.0=25_osxarm64_openblas
libbrotlicommon=1.1.0=hd74edd7_2
libbrotlidec=1.1.0=hd74edd7_2
libbrotlienc=1.1.0=hd74edd7_2
libcblas=3.9.0=25_osxarm64_openblas
libcxx=19.1.5=ha82da77_0
libdeflate=1.22=hd74edd7_0
libedit=3.1.20191231=hc8eb9b7_2
libexpat=2.6.4=h286801f_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=25_osxarm64_openblas
liblzma=5.6.3=h39f12f2_0
liblzma-devel=5.6.3=h39f12f2_0
libmpdec=4.0.0=h99b78c6_0
libopenblas=0.3.28=openmp_hf332438_1
libpng=1.6.44=hc14010f_0
libprotobuf=5.28.2=h8f0b736_0
libsodium=1.0.20=h99b78c6_0
libsqlite=3.47.0=hbaaea75_1
libtiff=4.7.0=hfce79cd_1
libtorch=2.5.1=cpu_generic_he23defe_5
libuv=1.49.2=h7ab814d_0
libwebp-base=1.4.0=h93a5062_0
libxcb=1.17.0=hdb1d25a_0
libzlib=1.3.1=h8359307_2
llvm-openmp=19.1.4=hdb05f8b_0
markupsafe=3.0.2=py312h998013c_1
matplotlib=3.9.3=py312h1f38498_0
matplotlib-base=3.9.3=py312hdbc7e53_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mistune=3.0.2=pyhd8ed1ab_1
mpc=1.3.1=h8f1351a_1
mpfr=4.2.1=hb693164_3
mpmath=1.3.0=pyhd8ed1ab_1
munkres=1.1.4=pyh9f0ad1d_0
nbclient=0.10.1=pyhd8ed1ab_0
nbconvert-core=7.16.4=pyhd8ed1ab_1
nbformat=5.10.4=pyhd8ed1ab_0
ncurses=6.5=h7bae524_1
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pyh267e887_2
nomkl=1.0=h5ca1d4c_0
notebook=7.3.0=pyhd8ed1ab_0
notebook-shim=0.2.4=pyhd8ed1ab_0
numpy=2.1.3=py312h94ee1e1_0
openjpeg=2.5.2=h9f1df11_0
openssl=3.4.0=h39f12f2_0
overrides=7.7.0=pyhd8ed1ab_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=py312hcd31e36_1
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.4=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.0.0=py312haf37ca6_0
pip=24.3.1=pyh8b19718_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_2
platformdirs=4.3.6=pyhd8ed1ab_1
prometheus_client=0.21.1=pyhd8ed1ab_0
prompt-toolkit=3.0.48=pyha770c72_1
prompt_toolkit=3.0.48=hd8ed1ab_1
psutil=6.1.0=py312h0bf5046_0
pthread-stubs=0.4=hd74edd7_1002
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_0
pycparser=2.22=pyh29332c3_1
pygments=2.18.0=pyhd8ed1ab_1
pyobjc-core=10.3.2=py312hb9d441b_0
pyobjc-framework-cocoa=10.3.2=py312hb9d441b_0
pyparsing=3.2.0=pyhd8ed1ab_2
pysocks=1.7.1=pyha55dd90_7
python=3.12.8=hc22306f_1_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python-fastjsonschema=2.21.1=pyhd8ed1ab_0
python-json-logger=2.0.7=pyhd8ed1ab_0
python-tzdata=2024.2=pyhd8ed1ab_1
python_abi=3.12=5_cp312
pytorch=2.5.1=cpu_generic_py312h73012d1_5
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.2=py312h024a12e_1
pyzmq=26.2.0=py312hf8a1cbd_3
qhull=2020.2=h420ef59_5
readline=8.2=h92ec313_1
referencing=0.35.1=pyhd8ed1ab_0
requests=2.32.3=pyhd8ed1ab_1
rfc3339-validator=0.1.4=pyhd8ed1ab_0
rfc3986-validator=0.1.1=pyh9f0ad1d_0
rpds-py=0.22.3=py312hcd83bfe_0
scikit-learn=1.5.2=py312h387f99c_1
scipy=1.14.1=py312h20deb59_1
send2trash=1.8.3=pyh31c8845_1
setuptools=75.6.0=pyhff2d567_1
six=1.16.0=pyhd8ed1ab_1
sleef=3.7=h8391f65_2
sniffio=1.3.1=pyhd8ed1ab_1
soupsieve=2.5=pyhd8ed1ab_1
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.13.3=pyh2585a3b_104
terminado=0.18.1=pyh31c8845_0
threadpoolctl=3.5.0=pyhc1e730c_0
tinycss2=1.4.0=pyhd8ed1ab_0
tk=8.6.13=h5083fa2_1
tomli=2.2.1=pyhd8ed1ab_1
torchvision=0.20.1=cpu_py312h2a91438_0
tornado=6.4.2=py312hea69d52_0
tqdm=4.67.1=pyhd8ed1ab_0
traitlets=5.14.3=pyhd8ed1ab_0
types-python-dateutil=2.9.0.20241003=pyhd8ed1ab_1
typing-extensions=4.12.2=hd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
typing_utils=0.1.0=pyhd8ed1ab_1
tzdata=2024b=hc8b5060_0
unicodedata2=15.1.0=py312h0bf5046_1
uri-template=1.3.0=pyhd8ed1ab_1
urllib3=2.2.3=pyhd8ed1ab_1
wcwidth=0.2.13=pyhd8ed1ab_1
webcolors=24.11.1=pyhd8ed1ab_0
webencodings=0.5.1=pyhd8ed1ab_3
websocket-client=1.8.0=pyhd8ed1ab_1
wheel=0.45.1=pyhd8ed1ab_1
widgetsnbextension=4.0.13=pyhd8ed1ab_1
xorg-libxau=1.0.11=hd74edd7_1
xorg-libxdmcp=1.1.5=hd74edd7_0
xz=5.6.3=h9a6d368_0
xz-gpl-tools=5.6.3=h9a6d368_0
xz-tools=5.6.3=h39f12f2_0
yaml=0.2.5=h3422bc3_2
zeromq=4.3.5=hc1bb282_7
zipp=3.21.0=pyhd8ed1ab_1
zstandard=0.23.0=py312h15fbf35_1
zstd=1.5.6=hb46c0d2_0
